+++ packages/react-grab/CHANGELOG.md
@@ -1,5 +1,17 @@
 # react-grab
 
+## 0.0.20
+
+### Patch Changes
+
+- fix: circular references issue
+
+## 0.0.19
+
+### Patch Changes
+
+- fix: react devtools and windows/linux compat
+
 ## 0.0.18
 
 ### Patch Changes